Introduce minimal test.
